+++ src/cluster.ts
@@ -1,14 +1,17 @@
-import { Cluster, PublicKey } from '@solana/web3.js'
+import { Cluster, clusterApiUrl, PublicKey } from '@solana/web3.js'
+
+export type PythCluster = Cluster | 'pythtest'
 
 /** Mapping from solana clusters to the public key of the pyth program. */
-const clusterToPythProgramKey: Record<Cluster, string> = {
+const clusterToPythProgramKey: Record<PythCluster, string> = {
   'mainnet-beta': 'FsJ3A3u2vn5cTVofAjvy6y5kwABJAqYWpe4975bi2epH',
   devnet: 'gSbePebfvPy7tRqimPoVecS2UsBvYv46ynrzWocc92s',
   testnet: '8tfDNiaEyrV6Q1U4DEXrEigs9DoDtkugzFbybENEbCDz',
+  pythtest: '8tfDNiaEyrV6Q1U4DEXrEigs9DoDtkugzFbybENEbCDz',
 }
 
 /** Gets the public key of the Pyth program running on the given cluster. */
-export function getPythProgramKeyForCluster(cluster: Cluster): PublicKey {
+export function getPythProgramKeyForCluster(cluster: PythCluster): PublicKey {
   if (clusterToPythProgramKey[cluster] !== undefined) {
     return new PublicKey(clusterToPythProgramKey[cluster])
   } else {
@@ -17,5 +20,15 @@ export function getPythProgramKeyForCluster(cluster: Cluster): PublicKey {
         Object.keys(clusterToPythProgramKey),
       )}`,
     )
+  }
+}
+
+/** Retrieves the RPC API URL for the specified Pyth cluster  */
+export function getPythClusterApiUrl(cluster: PythCluster): string {
+  //TODO: Add pythnet when it's ready
+  if (cluster === 'pythtest') {
+    return 'https://api.pythtest.pyth.network'
+  } else {
+    return clusterApiUrl(cluster)
   }
 }