+++ src/example_http_usage.ts
@@ -1,9 +1,9 @@
-import { Cluster, clusterApiUrl, Connection } from '@solana/web3.js'
-import { getPythProgramKeyForCluster } from './cluster'
+import { Connection } from '@solana/web3.js'
+import { getPythClusterApiUrl, getPythProgramKeyForCluster, PythCluster } from './cluster'
 import { PriceStatus, PythHttpClient } from '.'
 
-const SOLANA_CLUSTER_NAME: Cluster = 'mainnet-beta'
-const connection = new Connection(clusterApiUrl(SOLANA_CLUSTER_NAME))
+const SOLANA_CLUSTER_NAME: PythCluster = 'mainnet-beta'
+const connection = new Connection(getPythClusterApiUrl(SOLANA_CLUSTER_NAME))
 const pythPublicKey = getPythProgramKeyForCluster(SOLANA_CLUSTER_NAME)
 
 async function runQuery(): Promise<void> {